clean up term width detection and remove spurious warnings
Checking the terminal width is done in a number of places, including in
tests.  Outputting warnings whenever one of those routines is used on
something that isn't a terminal adds noise to tests and isn't helpful.
Eliminate the warnings for when using COLUMNS or falling back to 80.
Maintain the warning for when Term::Size::Any is not available.

We could also just add Term::Size::Any as a prerequisite.